filters/frame: use size_t for approx_size to avoid truncation
All users of this interface operate on size_t, and underlying structures
like AVBufferRef also express sizes in size_t.

Switch the approx size interface to size_t to avoid truncation and
match the expected type for byte sizes.

+++ audio/aframe.c
@@ -608,7 +608,7 @@ bool mp_aframe_reverse(struct mp_aframe *f)
     return true;
 }
 
-int mp_aframe_approx_byte_size(struct mp_aframe *frame)
+size_t mp_aframe_approx_byte_size(struct mp_aframe *frame)
 {
     // God damn, AVFrame is too fucking annoying. Just go with the size that
     // allocating a new frame would use.
